% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmtheme.R
\docType{data}
\name{mm_colors}
\alias{mm_colors}
\title{Adds color and theme options}
\format{An object of class \code{character} of length 6.}
\usage{
mm_colors
}
\description{
Adds color and theme options
}
\keyword{datasets}
