% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmtheme.R
\docType{data}
\name{mm_palettes}
\alias{mm_palettes}
\title{Define Color Palettes}
\format{An object of class \code{list} of length 5.}
\usage{
mm_palettes
}
\description{
Define Color Palettes
}
\keyword{datasets}
